several dependencies upgraded and use of deprecated APIs reduced: junit 4.8.2-> 4.10 dbunit 2.2 -> 2.4.8 jdom 1.0 -> 1.1 groovy-all 1.6.2 -> 1.8.6 commons-lang 2.4 -> 2.6 commons-io 1.4 -> 2.3 commons-beanutils 1.8.0 -> 1.8.3 commons-attributes-api 2.1 -> 2.2 commons-dbcp 1.2.2 -> 1.4 commons-net 1.4.1 -> 3.1 joda-time 1.6 -> 2.1 postgres 8.3-604.jdbc3 -> 9.1-901-1.jdbc4 freemarker 2.3.15 -> 2.3.19 xstream 1.3.1 -> 1.4.2 poi 3.7 -> 3.8 jaxen 1.1.1 -> 1.1.3 commons-codec 1.3 -> 1.6 xml-apis 1.0.b2 -> 2.0.2 jsontools-core 1.5 -> 1.7

+++ dbimport/src/test/java/com/agimatec/utility/fileimport/spreadsheet/ExcelImportTest.java
@@ -42,8 +42,8 @@ public class ExcelImportTest extends TestCase {
         HSSFWorkbook wb = new HSSFWorkbook(fs);
         HSSFSheet sheet = wb.getSheetAt(0);
         HSSFRow row = sheet.getRow(2);
-        HSSFCell cell = row.getCell((short) 3);
-        if (cell == null) cell = row.createCell((short) 3);
+        HSSFCell cell = row.getCell(3);
+        if (cell == null) cell = row.createCell(3);
         cell.setCellType(HSSFCell.CELL_TYPE_STRING);
         cell.setCellValue(new HSSFRichTextString("a test"));
